+++ packages/flutter_chat_view/lib/src/screens/chat_detail_screen.dart
@@ -53,7 +53,7 @@ class ChatDetailScreen extends StatefulWidget {
   final int pageSize;
   final double textfieldBottomPadding;
   final Color? iconDisabledColor;
-  final Function(String? userId) onPressUserProfile;
+  final Function(ChatUserModel user) onPressUserProfile;
   // ignore: avoid_positional_boolean_parameters
   final Widget? Function(BuildContext context)? loadingWidgetBuilder;
   final Widget Function(String userFullName)? usernameBuilder;